+++ cache/solidity-files-cache.json
@@ -1,0 +1,1 @@
+{"_format":"ethers-rs-sol-cache-3","paths":{"artifacts":"out","build_infos":"out/build-info","sources":"contracts","tests":"test","scripts":"script","libraries":["lib"]},"files":{"contracts/MotemaPool.sol":{"lastModificationDate":1710067592197,"contentHash":"218d7e38b46983b2f85981d05223e5e4","sourceName":"contracts/MotemaPool.sol","solcConfig":{"settings":{"optimizer":{"enabled":true,"runs":200},"metadata":{"useLiteralContent":false,"bytecodeHash":"ipfs","appendCBOR":true},"outputSelection":{"*":{"":["ast"],"*":["abi","evm.bytecode","evm.deployedBytecode","evm.methodIdentifiers","metadata"]}},"evmVersion":"paris","libraries":{}}},"imports":[],"versionRequirement":null,"artifacts":{"MotemaPool":{"0.8.24+commit.e11b9ed9.Darwin.appleclang":"MotemaPool.sol/MotemaPool.json"}}},"contracts/MotemaPool_flattened.sol":{"lastModificationDate":1710067590114,"contentHash":"1b2b985051a5c35be97517b457d375cd","sourceName":"contracts/MotemaPool_flattened.sol","solcConfig":{"settings":{"optimizer":{"enabled":true,"runs":200},"metadata":{"useLiteralContent":false,"bytecodeHash":"ipfs","appendCBOR":true},"outputSelection":{"*":{"":["ast"],"*":["abi","evm.bytecode","evm.deployedBytecode","evm.methodIdentifiers","metadata"]}},"evmVersion":"paris","libraries":{}}},"imports":[],"versionRequirement":"^0.8.0","artifacts":{"MotemaPool":{"0.8.24+commit.e11b9ed9.Darwin.appleclang":"MotemaPool_flattened.sol/MotemaPool.json"},"ReentrancyGuard":{"0.8.24+commit.e11b9ed9.Darwin.appleclang":"MotemaPool_flattened.sol/ReentrancyGuard.json"}}}}}
