+++ zkml_geiger/.env
@@ -1,2 +1,3 @@
 export MNEMONIC="frame mercy puppy spawn dinner true outside network neck spread vault local"
-export CONTRACT_ADDRESS="0x584dA09Cb8570074233812994646746b5D24e0FF"
+export CONTRACT_ADDRESS="0x8030A36c4a063752A0E2Ab9b66410c27E20B153B"
+export ALCHEMY_URL="https://eth-sepolia.g.alchemy.com/v2/aGP6ImVTfOlef4kzUxw0-31sabFgyTCT"
