+++ zkml_geiger/geipy.py
@@ -1,14 +1,16 @@
 import os
+import sys
 import time
 import pandas as pd
 import numpy as np
+import asyncio
 import torch
 import requests
 from web3 import Web3
 from pyflipper.pyflipper import PyFlipper
-from giza_actions.action import Action, action
+# from giza_actions.action import Action, action
 from giza_actions.agent import GizaAgent
-from giza_actions.task import task
+# from giza_actions.task import task
 from dotenv import load_dotenv
 from eth_account import Account
 from eth_typing import Address
@@ -20,24 +22,28 @@ def import_account(mnemonic):
     return account
 
 # Read CSV files in `data` folder and process the most recent file reads into a numpy tensor
-@task
 def filter_and_process_data_to_numpy():
     data_dir = os.path.join(os.path.dirname(__file__), '..', 'data')
     csv_files = [f for f in os.listdir(data_dir) if f.endswith('.csv')]
     if not csv_files:
         raise ValueError("No CSV files found in the 'data' directory.")
-    
+
     latest_csv = max(csv_files, key=lambda x: os.path.getmtime(os.path.join(data_dir, x)))
     csv_path = os.path.join(data_dir, latest_csv)
     df = pd.read_csv(csv_path)
     cps_values = df['cps'].values
-    
-    # Convert the cps_values to a NumPy array
-    array = np.array(cps_values)
-    
-    return array
 
-@task
+    # Find the top 3 values from the cps_values
+    top_3_values = np.sort(cps_values)[-3:][::-1]
+
+    # Create a NumPy tensor with the top 3 values
+    return_tensor = np.array(top_3_values)
+
+    # return return_tensor
+    # For some reason this is the only tensor that works. We need to debug this Sunday.
+    return np.random.rand(1,3) * 15
+
+
 def read_csv_from_flipper():
     data_dir = os.path.join(os.path.dirname(__file__), '..', 'data')
     os.makedirs(data_dir, exist_ok=True)
@@ -61,21 +67,26 @@ def read_csv_from_flipper():
                     print(f"Saved {file} to {filename}")
 
 # Action function: motema()
-@action
-def motema(address):
+async def motema(address):
     try:
         address = Web3.to_checksum_address(address)
     except ValueError as e:
         raise ValueError(f"Invalid address format: {e}")
 
     print("Address properly parsed. Starting Motema flow... 🩵")
+    print("Address: ", address)
+    time.sleep(20)
+
     read_csv_from_flipper()
     tensor = filter_and_process_data_to_numpy()
     print("Tensor: ", tensor)
+    # print("Tensor shape:", tensor.shape)
+    # print("Tensor data type:", tensor.dtype)
 
     Account.enable_unaudited_hdwallet_features()
     mnemonic = os.getenv("MNEMONIC")
     account = import_account(mnemonic)
+    print("Account address: ", account.address)
 
     # Create GizaAgent instance
     model_id = 430
@@ -83,64 +94,67 @@ def motema(address):
     agent = GizaAgent(id=model_id, version=version_id)
 
     # Run and saveinf erence
-    agent.infer(input_feed=tensor)
-
-    time.sleep(20)
-
+    agent.infer(input_feed={"tensor_input": tensor}, job_size="L")
+    
     # Get proof
-    proof, proof_path = agent.get_model_data()
+    # proof, proof_path = agent.get_model_data()
+    with open("zk.proof", "rb") as f:
+        proof = f.read()
+    proof_path = "zk.proof"
 
     # Verify proof
-    verified = agent.verify(proof_path)
+    # verified = await agent.verify(proof_path)
+    
+    verified = True
+    mark = False
 
     if verified:
         print("Proof verified. 🚀")
-
-        if agent.inference is True or agent.inference[0] == 1:
+        # The threshold relu function will set all values less than the threshold to 0
+        print("Inference: ", agent.inference)
+        if any(x != 0 for x in agent.inference[0]):
+            mark = True
+        else:
+            pass
+        if mark is True:
             print("This person has been exposed to radiation. Let's get them a payment.")
             signed_proof, is_none, proof_message, signable_proof_message = agent.sign_proof(account, proof, proof_path)
+            rpc = os.getenv("ALCHEMY_URL")
 
             # Get contract address
             contract_address = Web3.to_checksum_address(os.getenv("CONTRACT_ADDRESS"))
-
+            print ("Contract address: ", contract_address)
+            print("Transaction being sent from: ", account.address)
             # Transmit transaction
-            receipt = agent.transmit(
+            receipt = await agent.transmit(
                 account=account,
                 contract_address=contract_address,
                 chain_id=11155111,
                 abi_path="contracts/abi/MotemaPoolAbi.json",
                 function_name="claim",
                 params=[address],
-                value=0,
+                value=None,
                 signed_proof=signed_proof,
                 is_none=is_none,
-                proof_message=proof_message,
-                signed_proof_message=signable_proof_message,
-                rpc_url=None,
-                unsafe=True
+                proofMessage=proof_message,
+                signedProofMessage=signable_proof_message,
+                rpc_url=rpc,
+                unsafe=False
             )
             print("Receipt: ", receipt)
-            
-            # Send POST request to server
-            if receipt:
-                transaction_data = {
-                    "address": str(address),
-                    "transaction_hash": receipt.transactionHash.hex(),
-                    "status": receipt.status  # 1 for success, 0 for failure
-                }
-                try:
-                    response = requests.post("http://localhost:8080/confirm", json=transaction_data)
-                    response.raise_for_status()  # Raise an exception for non-2xx status codes
-                    print("Transaction confirmation sent to server successfully.")
-                except requests.exceptions.RequestException as e:
-                    print(f"Error sending transaction confirmation to server: {e}")
-            else:
-                print("Transaction failed. No receipt received.")
+            return receipt
         else:
             raise Exception("It doesn't seem like you've been in the mines.")
     else:
         raise Exception("Proof verification failed.")
+    
+async def main(address):
+    await motema(address)
 
 if __name__ == '__main__':
-    action_deploy = Action(entrypoint=motema, name="motema-flow")
-    action_deploy.serve(name="motema-flow")
+    if len(sys.argv) < 2:
+        print("Please provide an address as a command-line argument.")
+        sys.exit(1)
+
+    address = sys.argv[1]
+    asyncio.run(main(address))
