+++ zkml_geiger/templates/index.html
@@ -1,85 +1,73 @@
 <!DOCTYPE html>
 <html>
-  <head>
+<head>
     <title>Motema Flow</title>
     <style>
-      body {
-        font-family: Arial, sans-serif;
-        text-align: center;
-        padding-top: 50px;
-      }
-      input[type="text"] {
-        padding: 10px;
-        font-size: 16px;
-        width: 300px;
-      }
-      button {
-        padding: 10px 20px;
-        font-size: 16px;
-        background-color: #4CAF50;
-        color: white;
-        border: none;
-        cursor: pointer;
-      }
-      #transactionStatus {
-        margin-top: 20px;
-        font-weight: bold;
-      }
-      #transactionHash {
-        margin-top: 10px;
-        font-size: 14px;
-      }
+        body {
+            font-family: Arial, sans-serif;
+            text-align: center;
+            padding-top: 50px;
+        }
+
+        input[type="text"] {
+            padding: 10px;
+            font-size: 16px;
+            width: 300px;
+        }
+
+        button {
+            padding: 10px 20px;
+            font-size: 16px;
+            background-color: #4CAF50;
+            color: white;
+            border: none;
+            cursor: pointer;
+        }
+
+        #receipt {
+            margin-top: 20px;
+            font-weight: bold;
+        }
     </style>
-  </head>
-  <body>
+</head>
+<body>
     <h1>Motema Flow</h1>
     <p>Enter your Ethereum address and click "Start" to begin the flow:</p>
     <input type="text" id="addressInput" placeholder="Enter Ethereum address" />
     <br /><br />
     <button id="startButton">Start</button>
-    <div id="transactionStatus"></div>
-    <div id="transactionHash"></div>
+    <div id="receipt"></div>
 
     <script>
-      const startButton = document.getElementById("startButton");
-      const addressInput = document.getElementById("addressInput");
-      const transactionStatusDiv = document.getElementById("transactionStatus");
-      const transactionHashDiv = document.getElementById("transactionHash");
+        const startButton = document.getElementById("startButton");
+        const addressInput = document.getElementById("addressInput");
+        const receiptDiv = document.getElementById("receipt");
 
-      startButton.addEventListener("click", () => {
-        const address = addressInput.value;
-        if (address) {
-          fetch("/start", {
-            method: "POST",
-            headers: {
-              "Content-Type": "application/x-www-form-urlencoded",
-            },
-            body: `address=${encodeURIComponent(address)}`,
-          })
-            .then((response) => response.text())
-            .then((data) => {
-              alert(data);
-            })
-            .catch((error) => {
-              alert("An error occurred: " + error);
-            });
-        } else {
-          alert("Please enter an Ethereum address.");
-        }
-      });
-
-      function setupEventSource() {
-        const eventSource = new EventSource("/stream");
-        eventSource.onmessage = (event) => {
-          const data = JSON.parse(event.data);
-          console.log(data.message);
-        };
-        eventSource.onerror = (error) => {
-          console.error("EventSource failed:", error);
-        };
-      }
-
-      setupEventSource();
+        startButton.addEventListener("click", () => {
+            const address = addressInput.value;
+            if (address) {
+                fetch("/", {
+                    method: "POST",
+                    headers: {
+                        "Content-Type": "application/x-www-form-urlencoded",
+                    },
+                    body: `address=${encodeURIComponent(address)}`,
+                })
+                    .then((response) => response.json())
+                    .then((data) => {
+                        if (data.error) {
+                            receiptDiv.textContent = `Error: ${data.error}`;
+                        } else {
+                            receiptDiv.textContent = `Transaction Data: ${JSON.stringify(data)}`;
+                        }
+                    })
+                    .catch((error) => {
+                        receiptDiv.textContent = `Error: ${error}`;
+                    });
+            } else {
+                alert("Please enter an Ethereum address.");
+            }
+        });
     </script>
-  </body>
+</body>
 </html>
